Fixes microsoft/monaco-editor#2329: Move `registerThemingParticipant` call to `/editor/`
